Backend/requirements.txt
transformers
torch
passlib.context
fastapi
pydantic
uvicorn
venv